refactor(react): 拆出 Summary 与 Cancel 模块，引擎剥离总结/取消细节
把 ReactEngine 中总结阶段与取消处理的逻辑收敛到独立模块，主状态机只保留
ReAct 循环调度，不再直接承担 LLM 总结调用的细节。

新增模块：
- react/cancel.rs：取消原语（CancelStrategy / CancelSignal / emit_cancel_usage）。
  ReAct 主循环与总结阶段共享，按 Provider 协议区分取消时的 usage 处理策略。
- react/summary.rs：总结阶段（SummaryPhaseResult / run_summary_phase /
  parse_summary_need_more_work / strip_summary_marker）。run_summary_phase 以
  跨文件 impl 挂在 ReactEngine 上，force_final_response 仍留 context.rs，
  二者共同构成最终回复路径（后续按 Finalizer 进一步收敛）。

engine.rs 调整：
- TurnPhase / tools_for_current_turn 改为 pub(super)，供同级 summary 模块复用
- 移除已迁出的 run_summary_phase、parse_summary_need_more_work、
  strip_summary_marker 与不再使用的 request_for_summary_phase 导入

效果：engine.rs 从 2129 行降至 1782 行，总结与取消逻辑各自独立可测。
行为无变化，208 测试全通过。
